requirements.txt
aiogram
dotenv
openpyxl
pillow
qrcode
